refactor(scheduler): 引入 SchedulerContext trait 解耦定时任务与 Server
将执行器核心逻辑迁移到 tiangong-scheduler，通过 SchedulerContext trait
抽象消息发送和会话管理。Server 和 Desktop 各自实现 trait，Desktop 端
不再依赖 HTTP Server 即可执行定时任务。
