+++ crates/tiangong-core/src/scheduler/mod.rs
@@ -1,3 +1,4 @@
+pub use tiangong_scheduler::executor;
 pub use tiangong_scheduler::model;
 pub use tiangong_scheduler::store;
 pub use tiangong_scheduler::webhook;
