+++ crates/tiangong-server/Cargo.toml
@@ -9,6 +9,8 @@ tiangong-core = { workspace = true }
 tiangong-config = { workspace = true }
 tiangong-media = { workspace = true }
 tiangong-types = { workspace = true }
+tiangong-scheduler = { workspace = true }
+async-trait = { workspace = true }
 silent = { workspace = true, features = ["upgrade", "scheduler"] }
 tokio = { workspace = true, features = ["rt-multi-thread", "signal", "fs", "time"] }
 serde = { workspace = true, features = ["derive"] }
