+++ src-tauri/src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod app;
 pub mod commands;
+pub mod scheduler;
 pub mod view;
 
 pub use app::TiangongApp;
